Get rid of warning
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@1668 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/graphics/font.c
@@ -353,7 +353,7 @@ getCharFrame (FONT_DESC *fontPtr, wchar_t ch)
 	else
 	{
 #ifdef DEBUG
-		fprintf (stderr, "Character %u not present\n", ch);
+		fprintf (stderr, "Character %u not present\n", (unsigned int) ch);
 #endif
 		return NULL;
 	}
